requirements.txt
python-dotenv
aniso8601
astroid
blinker
certifi
chardet
click
flake8
Flask
Flask-HTTPAuth
Flask-Mail
Flask-RESTful
Flask-SQLAlchemy
Flask-Testing
gunicorn
idna
importlib-metadata
isort
itsdangerous
Jinja2
escape
lazy-object-proxy
MarkupSafe
mccabe
pycodestyle
pyflakes
PyJWT
pylint
pyperclip
python-dotenv
pytz
requests
selenium
six
SQLAlchemy
toml
urllib3
Werkzeug
wrapt
zipp
notebook
pandas
matplotlib
newspaper3k
autoscraper
torch
torchvision
torchaudio
transformers
sklearn
enelvo